requirements.txt
streamlit
pandas
numpy
plotly
seaborn
scikit-learn
python-dotenv
